Improve rpdTracerControl usage

+++ rpd_tracer/rpdTracerControl.py
@@ -41,10 +41,11 @@ class rpdTracerControl:
     __filename = "trace.rpd"
     __rpd = None    # the dll/
     __initFile = True
+    __active = True
 
     @classmethod
     def loadLibrary(cls):
-        if cls.__rpd == None:
+        if cls.__rpd == None and cls.__active == True:
             os.environ["RPDT_AUTOSTART"] = "0"
             cls.__rpd = CDLL(find_library("rpd_tracer"))
 
@@ -53,8 +54,7 @@ class rpdTracerControl:
         # Only the top parent process will initialize the trace file
         if isChildProcess() or os.getenv("RPDT_FILENAME"):
             cls.__initFile = False
-
-        if cls.__initFile == True:
+        if cls.__initFile == True and cls.__active == True:
             if os.path.exists(cls.__filename):
                 os.remove(cls.__filename)
             # Create new file and write schema
@@ -64,9 +64,9 @@ class rpdTracerControl:
             connection.commit()
             connection.close()
             #
-            os.environ["RPDT_FILENAME"] = cls.__filename
+            #os.environ["RPDT_FILENAME"] = cls.__filename
             cls.__initFile = False   
-
+        os.environ["RPDT_FILENAME"] = cls.__filename
 
     # You can set the output filename and optionally append to an exiting file.
     #   This must be done on the main process before any class instances are created.
@@ -79,7 +79,8 @@ class rpdTracerControl:
     def setFilename(cls, name, append = False):
         if os.getenv("RPDT_FILENAME"):
             cls.__filename = os.getenv("RPDT_FILENAME")
-            raise Warning(f"RPDT_FILENAME is already initialized. Logging into {cls.__filename}")
+            if name != cls.__filename:
+                raise Warning(f"RPDT_FILENAME is already initialized. Logging into {cls.__filename}")
             return
 
         if cls.__rpd != None:
@@ -89,9 +90,19 @@ class rpdTracerControl:
 
 
         cls.__filename = name
+        #os.environ["RPDT_FILENAME"] = cls.__filename
         if append:
-            os.environ["RPDT_FILENAME"] = cls.__filename
             cls.__initFile = False
+
+    @classmethod
+    def skipCreate(cls):
+        cls.__initFile = False
+
+    @classmethod
+    def skipLoad(cls):
+       # If the tracer is not already loaded, then everything is a no-op.
+       if os.environ.get("RPDT_LOADED") == None:
+           cls.__active = False
 
     def __init__(self):
         rpdTracerControl.initializeFile()
@@ -101,13 +112,16 @@ class rpdTracerControl:
         pass
 
     def start(self):
-        rpdTracerControl.__rpd.rpdstart()
+        if rpdTracerControl.__rpd:
+            rpdTracerControl.__rpd.rpdstart()
 
     def stop(self):
-        rpdTracerControl.__rpd.rpdstop()
+        if rpdTracerControl.__rpd:
+            rpdTracerControl.__rpd.rpdstop()
 
     def flush(self):
-        rpdTracerControl.__rpd.rpdflush()
+        if rpdTracerControl.__rpd:
+            rpdTracerControl.__rpd.rpdflush()
 
     def __enter__(self):
         self.start()
@@ -116,10 +130,12 @@ class rpdTracerControl:
         self.stop()
 
     def rangePush(self, domain: str, apiName: str, args: str):
-        rpdTracerControl.__rpd.rpd_rangePush(bytes(domain, encoding='utf-8'), bytes(apiName, encoding='utf-8'), bytes(args, encoding='utf-8'))
+        if rpdTracerControl.__rpd:
+            rpdTracerControl.__rpd.rpd_rangePush(bytes(domain, encoding='utf-8'), bytes(apiName, encoding='utf-8'), bytes(args, encoding='utf-8'))
 
     def rangePop(self):
-        rpdTracerControl.__rpd.rpd_rangePop()
+        if rpdTracerControl.__rpd:
+            rpdTracerControl.__rpd.rpd_rangePop()
 
 
     # python stack tracing
